Move old ProductsCard to temp backup file and add resource picker to the Products Card
